third_party/WebKit/LayoutTests/http/tests/navigation/rename-subframe-goback-expected.txt

Tests that session history works even if the page changes frame names after initial load.




--------
Frame: 'initial-name'
--------
This page was requested with the HTTP method POST.

Parameters:

the-input = input value goes here
submitwithpost = Submit with POST
Http headers:

HTTP_CACHE_CONTROL = max-age=0
HTTP_CONNECTION = keep-alive
HTTP_HOST = 127.0.0.1:8000
HTTP_ORIGIN = http://127.0.0.1:8000
HTTP_REFERER = http://127.0.0.1:8000/navigation/resources/page-that-posts.html
HTTP_SEC_METADATA = cause=forced, destination=nested-document, site=same-origin
HTTP_UPGRADE_INSECURE_REQUESTS = 1
